% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deq_pipeline.R
\name{deq}
\alias{deq}
\title{DEQ}
\usage{
deq(input.bams, ip.bams, treated.input.bams, treated.ip.bams, peak.files,
  gtf, paired.end = FALSE, outfi = "deq_results.txt", tool = "deq",
  compare.gene = TRUE, readlen = 100, fraglen = 100, nthreads = 1)
}
\arguments{
\item{input.bams}{vector of control input bam files}

\item{ip.bams}{vector of control IP bam files}

\item{peak.files}{vector of peak file names}

\item{gtf}{location of a gtf file for annotations}

\item{paired.end}{boolean indicating whether reads were paired end (defaults
to FALSE)}

\item{outfi}{output file name (defaults to deq.results.txt)}

\item{tool}{which tools to run (options: any combination of d, e, and q;
defaults to all: 'deq')}

\item{compare.gene}{whether to calculate changes in gene expression for
comparison (highly recommended, defaults to TRUE)}

\item{readlen}{average read length for the data (defaults to 100)}

\item{fraglen}{average fragment length for the data (defaults to 100) --
(fraglen-readlen) is used to calculate the extension of reads for counting}

\item{nthreads}{number of threads to run on (defaults to 1)}

\item{input.bams}{vector of treatment input bam files}

\item{ip.bams}{vector of treatment IP bam files}
}
\value{
GRanges object with peaks and estimated changes
}
\description{
A package to conveniently run DESeq2, edgeR, and QNB for the detection of differential methylation in MeRIP/m6A-seq data.
}
